dbeaver/pro#3763 Fix UI in data transfer
