+++ plugins/org.jkiss.dbeaver.ui.navigator/src/org/jkiss/dbeaver/ui/internal/UINavigatorMessages.properties
@@ -206,8 +206,6 @@ pref_page_target_link_show_datasource_settings = Datasource settings
 pref_page_target_link_show_global_settings = Global settings
 
 label_active_service_instance=Active server instance
-label_choose=Choose {0}
-label_catalog_schema=Catalog/Schema
 label_error_list=Error loading instance list
 label_instance=Instance
