+++ plugins/org.jkiss.dbeaver.ui.navigator/src/org/jkiss/dbeaver/ui/internal/UINavigatorMessages_es.properties
@@ -147,9 +147,6 @@ error_sql_generation_title = Error de generaci\u00F3n de script
 
 label_active_service_instance = Instancia de servidor activa
 
-label_choose = Elegir {0}
-label_catalog_schema = Cat\u00E1logo/Esquema
-
 label_configure_columns = Configurar columnas
 
 label_description = Descripci\u00F3n
