+++ plugins/org.jkiss.dbeaver.ui.navigator/src/org/jkiss/dbeaver/ui/internal/UINavigatorMessages_it.properties
@@ -119,9 +119,6 @@ editors_entity_properties_text = Propriet\u00E0
 
 editors_entity_properties_tooltip_suffix = \ Propriet\u00E0
 
-label_choose = Scegli {0}
-label_catalog_schema = Catalogo/Schema
-
 label_configure_columns = Configura le colonne
 
 label_description = Descrizione
