+++ plugins/org.jkiss.dbeaver.ui.navigator/src/org/jkiss/dbeaver/ui/internal/UINavigatorMessages_ja.properties
@@ -159,8 +159,6 @@ pref_page_target_button_use_datasource_settings =\u30C7\u30FC\u30BF\u30BD\u30FC\
 pref_page_target_link_show_datasource_settings =\u30C7\u30FC\u30BF\u30BD\u30FC\u30B9\u8A2D\u5B9A
 pref_page_target_link_show_global_settings =\u5168\u4F53\u8A2D\u5B9A
 
-label_choose = {0} \u9078\u629E
-label_catalog_schema = \u30AB\u30BF\u30ED\u30B0/\u30B9\u30AD\u30FC\u30DE\u3092
 label_error_list = \u30A4\u30F3\u30B9\u30BF\u30F3\u30B9\u306E\u8AAD\u307F\u8FBC\u307F\u30A8\u30E9\u30FC
 label_instance = \u30A4\u30F3\u30B9\u30BF\u30F3\u30B9
