+++ plugins/org.jkiss.dbeaver.ui.navigator/src/org/jkiss/dbeaver/ui/internal/UINavigatorMessages_pt_BR.properties
@@ -221,9 +221,6 @@ filter_connection_name_placeholder = Informe parte do nome da conex\u00E3o aqui
 
 label_active_service_instance = Inst\u00E2ncia de servidor ativa
 
-label_choose = Selecionar {0}
-label_catalog_schema =  Cat\u00E1logo/Esquema
-
 label_configure_columns = Configurar colunas
 
 label_description = Descri\u00E7\u00E3o
