+++ plugins/org.jkiss.dbeaver.ui.navigator/src/org/jkiss/dbeaver/ui/internal/UINavigatorMessages_ro.properties
@@ -208,8 +208,6 @@ pref_page_target_link_show_datasource_settings = Set\u0103ri surse de date
 pref_page_target_link_show_global_settings = Set\u0103ri globale
 
 label_active_service_instance = Instan\u0163\u0103 de server activ
-label_choose = Alege\u0163i {0}
-label_catalog_schema = Catalogul/Schema
 
 label_error_list = Eroare la \u00EEnc\u0103rcarea listei de instan\u0163e
 label_instance = Instan\u0163\u0103
