+++ plugins/org.jkiss.dbeaver.ui.navigator/src/org/jkiss/dbeaver/ui/internal/UINavigatorMessages_ru.properties
@@ -156,8 +156,6 @@ pref_page_target_link_show_global_settings=\u0413\u043B\u043E\u0431\u0430\u043B\
 ui_navigator_loading_text_loading=\u0417\u0430\u0433\u0440\u0443\u0437\u043A\u0430
 
 label_active_service_instance=\u042D\u043A\u0437\u0435\u043C\u043F\u043B\u044F\u0440 \u0430\u043A\u0442\u0438\u0432\u043D\u043E\u0433\u043E \u0441\u0435\u0440\u0432\u0435\u0440\u0430
-label_choose=\u0412\u044B\u0431\u0440\u0430\u0442\u044C {0}
-label_catalog_schema = \u041A\u0430\u0442\u0430\u043B\u043E\u0433/\u0421\u0445\u0435\u043C\u0430
 label_error_list=\u041E\u0448\u0438\u0431\u043A\u0430 \u0437\u0430\u0433\u0440\u0443\u0437\u043A\u0438 \u0441\u043F\u0438\u0441\u043A\u0430 \u044D\u043A\u0437\u0435\u043C\u043F\u043B\u044F\u0440\u043E\u0432
 label_instance=\u042D\u043A\u0437\u0435\u043C\u043F\u043B\u044F\u0440
