+++ plugins/org.jkiss.dbeaver.ui.navigator/src/org/jkiss/dbeaver/ui/internal/UINavigatorMessages_uk.properties
@@ -207,8 +207,6 @@ pref_page_target_link_show_datasource_settings = \u041D\u0430\u043B\u0430\u0448\
 pref_page_target_link_show_global_settings = \u0413\u043B\u043E\u0431\u0430\u043B\u044C\u043D\u0456 \u043D\u0430\u043B\u0430\u0448\u0442\u0443\u0432\u0430\u043D\u043D\u044F
 
 label_active_service_instance=\u0410\u043A\u0442\u0438\u0432\u043D\u0438\u0439 \u0435\u043A\u0437\u0435\u043C\u043F\u043B\u044F\u0440 \u0441\u0435\u0440\u0432\u0435\u0440\u0430
-label_choose=\u0412\u0438\u0431\u0440\u0430\u0442\u0438 {0}
-label_catalog_schema = \u041A\u0430\u0442\u0430\u043B\u043E\u0433/\u0441\u0445\u0435\u043C\u0430
 label_error_list=\u041F\u043E\u043C\u0438\u043B\u043A\u0430 \u0437\u0430\u0432\u0430\u043D\u0442\u0430\u0436\u0435\u043D\u043D\u044F \u0441\u043F\u0438\u0441\u043A\u0443 \u0435\u043A\u0437\u0435\u043C\u043F\u043B\u044F\u0440\u0456\u0432
 label_instance=\u0415\u043A\u0437\u0435\u043C\u043F\u043B\u044F\u0440
