+++ plugins/org.jkiss.dbeaver.ui.navigator/src/org/jkiss/dbeaver/ui/internal/UINavigatorMessages_zh.properties
@@ -221,9 +221,6 @@ filter_connection_name_placeholder = \u5728\u6B64\u5904\u8F93\u5165\u8FDE\u63A5\
 
 label_active_service_instance = \u6D3B\u52A8\u670D\u52A1\u5668\u5B9E\u4F8B
 
-label_choose = \u9009\u62E9 {0}
-label_catalog_schema = \u76EE\u5F55/\u6A21\u5F0F
-
 label_configure_columns = \u914D\u7F6E\u5217
 
 label_description = \u63CF\u8FF0
