+++ plugins/org.jkiss.dbeaver.ui/src/org/jkiss/dbeaver/ui/internal/UIMessages.properties
@@ -96,3 +96,5 @@ preference_page_no_access = No access to this preference page, you are missing t
 link_external_label = {0} \uD83E\uDC55
 link_external_tip = Show in global preferences
 
+label_catalog_schema=Catalog/Schema
+label_choose=Choose {0}
