+++ plugins/org.jkiss.dbeaver.ui/src/org/jkiss/dbeaver/ui/internal/UIMessages_es.properties
@@ -5,3 +5,5 @@ controls_locale_selector_label_locale = Localizaci\u00F3n
 controls_locale_selector_label_variant = Variante
 
 control_label_connection_folder = Carpeta de conexi\u00F3n
+label_catalog_schema = Cat\u00E1logo/Esquema
+label_choose = Elegir {0}
