+++ plugins/org.jkiss.dbeaver.ui/src/org/jkiss/dbeaver/ui/internal/UIMessages_it.properties
@@ -114,3 +114,5 @@ ui_properties_tree_viewer_action_reset_value = Reimposta valore
 ui_properties_tree_viewer_category_general = Generale
 
 ui_properties_value = Valore
+label_catalog_schema = Catalogo/Schema
+label_choose = Scegli {0}
