+++ plugins/org.jkiss.dbeaver.ui/src/org/jkiss/dbeaver/ui/internal/UIMessages_ja.properties
@@ -39,3 +39,5 @@ ui_properties_name =\u540D
 ui_properties_value =\u5024
 
 control_label_connection_folder =\u63A5\u7D9A\u30D5\u30A9\u30EB\u30C0
+label_catalog_schema = \u30AB\u30BF\u30ED\u30B0/\u30B9\u30AD\u30FC\u30DE\u3092
+label_choose = {0} \u9078\u629E
