+++ plugins/org.jkiss.dbeaver.ui/src/org/jkiss/dbeaver/ui/internal/UIMessages_pt_BR.properties
@@ -148,3 +148,5 @@ ui_properties_value = Valor
 utils_actions_copy_all_label = Copiar tudo
 
 utils_actions_copy_label = Copiar {0}
+label_catalog_schema =  Cat\u00E1logo/Esquema
+label_choose = Selecionar {0}
