+++ plugins/org.jkiss.dbeaver.ui/src/org/jkiss/dbeaver/ui/internal/UIMessages_ro.properties
@@ -92,3 +92,5 @@ text_with_open_dialog_browse_remote = R\u0103sfoi\u0163i sistemul de fi\u015Fier
 text_with_open_dialog_edit_file = Edita\u0163i fi\u015Fierul
 
 notification_popup_context_message = Notificare {0}
+label_catalog_schema = Catalogul/Schema
+label_choose = Alege\u0163i {0}
