+++ plugins/org.jkiss.dbeaver.ui/src/org/jkiss/dbeaver/ui/internal/UIMessages_ru.properties
@@ -66,3 +66,5 @@ control_font_italic = \u041A\u0443\u0440\u0441\u0438\u0432
 control_font_bold = \u0416\u0438\u0440\u043D\u044B\u0439
 
 control_default_color_selector_reset_default_tip = \u0418\u0441\u043F\u043E\u043B\u044C\u0437\u0443\u0439\u0442\u0435 \u043F\u0440\u0430\u0432\u0443\u044E \u043A\u043D\u043E\u043F\u043A\u0443 \u043C\u044B\u0448\u0438 \u0434\u043B\u044F \u0441\u0431\u0440\u043E\u0441\u0430 \u0446\u0432\u0435\u0442\u0430 \u0434\u043E \u0437\u043D\u0430\u0447\u0435\u043D\u0438\u044F \u043F\u043E \u0443\u043C\u043E\u043B\u0447\u0430\u043D\u0438\u044E
+label_catalog_schema = \u041A\u0430\u0442\u0430\u043B\u043E\u0433/\u0421\u0445\u0435\u043C\u0430
+label_choose=\u0412\u044B\u0431\u0440\u0430\u0442\u044C {0}
