+++ plugins/org.jkiss.dbeaver.ui/src/org/jkiss/dbeaver/ui/internal/UIMessages_uk.properties
@@ -89,3 +89,5 @@ text_with_open_dialog_edit_text = \u0420\u0435\u0434\u0430\u0433\u0443\u0432\u04
 text_with_open_dialog_browse = \u041E\u0433\u043B\u044F\u0434...
 text_with_open_dialog_browse_remote = \u041E\u0433\u043B\u044F\u0434 \u0432\u0456\u0434\u0434\u0430\u043B\u0435\u043D\u043E\u0457 \u0444\u0430\u0439\u043B\u043E\u0432\u043E\u0457 \u0441\u0438\u0441\u0442\u0435\u043C\u0438
 text_with_open_dialog_edit_file = \u0420\u0435\u0434\u0430\u0433\u0443\u0432\u0430\u0442\u0438 \u0444\u0430\u0439\u043B
+label_catalog_schema = \u041A\u0430\u0442\u0430\u043B\u043E\u0433/\u0441\u0445\u0435\u043C\u0430
+label_choose=\u0412\u0438\u0431\u0440\u0430\u0442\u0438 {0}
