+++ plugins/org.jkiss.dbeaver.ui/src/org/jkiss/dbeaver/ui/internal/UIMessages_zh.properties
@@ -132,3 +132,5 @@ ui_properties_value = \u503C
 utils_actions_copy_all_label = \u590D\u5236\u5168\u90E8
 
 utils_actions_copy_label = \u590D\u5236 {0}
+label_catalog_schema = \u76EE\u5F55/\u6A21\u5F0F
+label_choose = \u9009\u62E9 {0}
